+++ modules/core/dependency/python-ihm/test/test_reader.py
@@ -1242,10 +1242,11 @@ _ihm_ensemble_info.num_ensemble_models_deposited
 _ihm_ensemble_info.ensemble_precision_value
 _ihm_ensemble_info.ensemble_file_id
 1 'Cluster 1' 2 3 . dRMSD 1257 1 15.400 9
+2 'Cluster 2' 2 . . dRMSD 1257 1 15.400 9
 """
         for fh in cif_file_handles(cif):
             s, = ihm.reader.read(fh)
-            e, = s.ensembles
+            e, e2 = s.ensembles
             self.assertEqual(e.model_group._id, '3')
             self.assertEqual(e.num_models, 1257)
             self.assertEqual(e.post_process._id, '2')
@@ -1254,6 +1255,7 @@ _ihm_ensemble_info.ensemble_file_id
             self.assertEqual(e.name, 'Cluster 1')
             self.assertAlmostEqual(e.precision, 15.4, places=1)
             self.assertEqual(e.file._id, '9')
+            self.assertIsNone(e2.model_group)
 
     def test_density_handler(self):
         """Test DensityHandler"""