+++ modules/core/dependency/python-ihm/test/test_restraint.py
@@ -154,6 +154,10 @@ class Tests(unittest.TestCase):
         f.ranges.append(a(2,2))
         self.assertEqual(f.type, 'residue')
 
+        # Should work with actual Residue objects too
+        f.ranges.append(a.residue(3))
+        self.assertEqual(f.type, 'residue')
+
         # At least one range is a true range
         f.ranges.append(a(3,4))
         self.assertEqual(f.type, 'residue range')
@@ -194,6 +198,8 @@ class Tests(unittest.TestCase):
         self.assertEqual(f.dataset, 'foo')
         self.assertEqual(f.object_characteristic, 'outer surface')
         self.assertIsNone(f.assembly)
+        self.assertEqual(f.feature, 'feat')
+        self.assertEqual(f._all_features, ('feat',))
 
     def test_derived_distance_restraint(self):
         """Test DerivedDistanceRestraint class"""
@@ -202,6 +208,9 @@ class Tests(unittest.TestCase):
                 distance='dist')
         self.assertEqual(f.dataset, 'foo')
         self.assertIsNone(f.assembly)
+        self.assertEqual(f.feature1, 'feat1')
+        self.assertEqual(f.feature2, 'feat2')
+        self.assertEqual(f._all_features, ('feat1', 'feat2'))
 
 
 if __name__ == '__main__':